share metadata between the Catalog and the 'iceberg_scan'
