+++ src/catalog_api.cpp
@@ -11,9 +11,6 @@
 #include "duckdb/common/http_util.hpp"
 #include "duckdb/common/exception/http_exception.hpp"
 
-#include "storage/authorization/sigv4.hpp"
-#include "storage/authorization/oauth2.hpp"
-
 #include "rest_catalog/objects/list.hpp"
 
 using namespace duckdb_yyjson;
@@ -74,150 +71,6 @@ static string GetTableMetadataCached(ClientContext &context, IRCatalog &catalog,
 
 vector<string> IRCAPI::GetCatalogs(ClientContext &context, IRCatalog &catalog) {
 	throw NotImplementedException("ICAPI::GetCatalogs");
-}
-
-static void ParseConfigOptions(const case_insensitive_map_t<string> &config, case_insensitive_map_t<Value> &options) {
-	//! Set of recognized config parameters and the duckdb secret option that matches it.
-	static const case_insensitive_map_t<string> config_to_option = {{"s3.access-key-id", "key_id"},
-	                                                                {"s3.secret-access-key", "secret"},
-	                                                                {"s3.session-token", "session_token"},
-	                                                                {"s3.region", "region"},
-	                                                                {"region", "region"},
-	                                                                {"client.region", "region"},
-	                                                                {"s3.endpoint", "endpoint"}};
-
-	if (config.empty()) {
-		return;
-	}
-	for (auto &entry : config) {
-		auto it = config_to_option.find(entry.first);
-		if (it != config_to_option.end()) {
-			options[it->second] = entry.second;
-		}
-	}
-
-	auto it = config.find("s3.path-style-access");
-	if (it != config.end()) {
-		bool path_style;
-		if (it->second == "true") {
-			path_style = true;
-		} else if (it->second == "false") {
-			path_style = false;
-		} else {
-			throw InvalidInputException("Unexpected value ('%s') for 's3.path-style-access' in 'config' property",
-			                            it->second);
-		}
-
-		options["use_ssl"] = Value(!path_style);
-		if (path_style) {
-			options["url_style"] = "path";
-		}
-	}
-
-	auto endpoint_it = options.find("endpoint");
-	if (endpoint_it == options.end()) {
-		return;
-	}
-	auto endpoint = endpoint_it->second.ToString();
-	if (StringUtil::StartsWith(endpoint, "http://")) {
-		endpoint = endpoint.substr(7, string::npos);
-	}
-	if (StringUtil::StartsWith(endpoint, "https://")) {
-		endpoint = endpoint.substr(8, string::npos);
-	}
-	if (StringUtil::EndsWith(endpoint, "/")) {
-		endpoint = endpoint.substr(0, endpoint.size() - 1);
-	}
-	endpoint_it->second = endpoint;
-}
-
-IRCAPITableCredentials IRCAPI::GetTableCredentials(ClientContext &context, IRCatalog &catalog, const string &schema,
-                                                   const string &table, const string &secret_base_name) {
-	IRCAPITableCredentials result;
-	string api_result = GetTableMetadataCached(context, catalog, schema, table);
-	std::unique_ptr<yyjson_doc, YyjsonDocDeleter> doc(ICUtils::api_result_to_doc(api_result));
-	auto *root = yyjson_doc_get_root(doc.get());
-	auto load_table_result = rest_api_objects::LoadTableResult::FromJSON(root);
-
-	case_insensitive_map_t<Value> user_defaults;
-	if (catalog.auth_handler->type == IRCAuthorizationType::SIGV4) {
-		auto &sigv4_auth = catalog.auth_handler->Cast<SIGV4Authorization>();
-		auto catalog_credentials = IRCatalog::GetStorageSecret(context, sigv4_auth.secret);
-		// start with the credentials needed for the catalog and overwrite information contained
-		// in the vended credentials. We do it this way to maintain the region info from the catalog credentials
-		if (catalog_credentials) {
-			auto kv_secret = dynamic_cast<const KeyValueSecret &>(*catalog_credentials->secret);
-			for (auto &option : kv_secret.secret_map) {
-				// Ignore refresh info.
-				// if the credentials are the same as for the catalog, then refreshing the catalog secret is enough
-				// otherwise the vended credentials contain their own information for refreshing.
-				if (option.first != "refresh_info" && option.first != "refresh") {
-					user_defaults.emplace(option);
-				}
-			}
-		}
-	} else if (catalog.auth_handler->type == IRCAuthorizationType::OAUTH2) {
-		auto &oauth2_auth = catalog.auth_handler->Cast<OAuth2Authorization>();
-		if (!oauth2_auth.default_region.empty()) {
-			user_defaults["region"] = oauth2_auth.default_region;
-		}
-	}
-
-	// Mapping from config key to a duckdb secret option
-
-	case_insensitive_map_t<Value> config_options;
-	//! TODO: apply the 'defaults' retrieved from the /v1/config endpoint
-	config_options.insert(user_defaults.begin(), user_defaults.end());
-
-	if (load_table_result.has_config) {
-		auto &config = load_table_result.config;
-		ParseConfigOptions(config, config_options);
-	}
-
-	const auto &metadata_location = load_table_result.metadata.location;
-
-	if (load_table_result.has_storage_credentials) {
-		auto &storage_credentials = load_table_result.storage_credentials;
-
-		//! If there is only one credential listed, we don't really care about the prefix,
-		//! we can use the metadata_location instead.
-		const bool ignore_credential_prefix = storage_credentials.size() == 1;
-		for (idx_t index = 0; index < storage_credentials.size(); index++) {
-			auto &credential = storage_credentials[index];
-			CreateSecretInput create_secret_input;
-			create_secret_input.on_conflict = OnCreateConflict::REPLACE_ON_CONFLICT;
-			create_secret_input.persist_type = SecretPersistType::TEMPORARY;
-
-			create_secret_input.scope.push_back(ignore_credential_prefix ? metadata_location : credential.prefix);
-			create_secret_input.name = StringUtil::Format("%s_%d_%s", secret_base_name, index, credential.prefix);
-
-			create_secret_input.type = "s3";
-			create_secret_input.provider = "config";
-			create_secret_input.storage_type = "memory";
-			create_secret_input.options = config_options;
-
-			ParseConfigOptions(credential.config, create_secret_input.options);
-			//! TODO: apply the 'overrides' retrieved from the /v1/config endpoint
-			result.storage_credentials.push_back(create_secret_input);
-		}
-	}
-
-	if (result.storage_credentials.empty() && !config_options.empty()) {
-		//! Only create a secret out of the 'config' if there are no 'storage-credentials'
-		result.config = make_uniq<CreateSecretInput>();
-		auto &config = *result.config;
-		config.on_conflict = OnCreateConflict::REPLACE_ON_CONFLICT;
-		config.persist_type = SecretPersistType::TEMPORARY;
-
-		//! TODO: apply the 'overrides' retrieved from the /v1/config endpoint
-		config.options = config_options;
-		config.name = secret_base_name;
-		config.type = "s3";
-		config.provider = "config";
-		config.storage_type = "memory";
-	}
-
-	return result;
 }
 
 rest_api_objects::LoadTableResult IRCAPI::GetTable(ClientContext &context, IRCatalog &catalog, const string &schema,
