+++ src/include/catalog_api.hpp
@@ -18,17 +18,9 @@ struct IRCAPISchema {
 	string catalog_name;
 };
 
-struct IRCAPITableCredentials {
-	unique_ptr<CreateSecretInput> config;
-	vector<CreateSecretInput> storage_credentials;
-};
-
 class IRCAPI {
 public:
 	static const string API_VERSION_1;
-
-	static IRCAPITableCredentials GetTableCredentials(ClientContext &context, IRCatalog &catalog, const string &schema,
-	                                                  const string &table, const string &secret_base_name);
 	static vector<string> GetCatalogs(ClientContext &context, IRCatalog &catalog);
 	static vector<rest_api_objects::TableIdentifier> GetTables(ClientContext &context, IRCatalog &catalog,
 	                                                           const string &schema);
