+++ src/include/iceberg_metadata.hpp
@@ -43,11 +43,19 @@ namespace duckdb {
 
 struct IcebergScanInfo : public TableFunctionInfo {
 public:
-	IcebergScanInfo(IcebergTableMetadata &metadata, optional_ptr<IcebergSnapshot> snapshot, IcebergTableSchema &schema)
-	    : metadata(metadata), snapshot(snapshot), schema(schema) {
+	IcebergScanInfo(const string &metadata_path, IcebergTableMetadata &metadata, optional_ptr<IcebergSnapshot> snapshot,
+	                IcebergTableSchema &schema)
+	    : metadata_path(metadata_path), metadata(metadata), snapshot(snapshot), schema(schema) {
+	}
+	IcebergScanInfo(const string &metadata_path, unique_ptr<IcebergTableMetadata> owned_metadata_p,
+	                optional_ptr<IcebergSnapshot> snapshot, IcebergTableSchema &schema)
+	    : metadata_path(metadata_path), owned_metadata(std::move(owned_metadata_p)), metadata(*owned_metadata),
+	      snapshot(snapshot), schema(schema) {
 	}
 
 public:
+	string metadata_path;
+	unique_ptr<IcebergTableMetadata> owned_metadata;
 	IcebergTableMetadata &metadata;
 	optional_ptr<IcebergSnapshot> snapshot;
 	IcebergTableSchema &schema;
